Remove http middleware in and out
